Rename: 프로필 관련 페이지 구조 수정
