+++ src/pages/index.ts
@@ -2,3 +2,4 @@ export * from './group'
 export * from './group-management'
 export * from './point'
 export * from './auth'
+export * from './profile'
