+++ src/pages/profile/default/components/index.ts
@@ -1,0 +1,5 @@
+export * from './navigation'
+export * from './my-profile'
+export * from './user-profile'
+export * from './ranking'
+export * from './oven-menu'
