+++ src/pages/profile/default/components/my-profile/edit-profile-form/index.ts
@@ -1,0 +1,1 @@
+export { EditProfileForm } from './EditProfileForm'
