+++ src/pages/profile/default/components/my-profile/index.ts
@@ -1,0 +1,1 @@
+export { MyProfile } from './MyProfile'
