+++ src/pages/profile/default/components/my-profile/profile-image/index.ts
@@ -1,0 +1,1 @@
+export { ProfileImage } from './ProfileImage'
