+++ src/pages/profile/default/components/my-profile/visited-tag/index.ts
@@ -1,0 +1,1 @@
+export { VisitedTag } from './VisitedTag'
