+++ src/pages/profile/default/components/navigation/index.ts
@@ -1,0 +1,1 @@
+export { ProfileNavigation } from './Navigation'
