+++ src/pages/profile/default/components/oven-menu/index.ts
@@ -1,0 +1,1 @@
+export { OvenMenu } from './OvenMenu'
