+++ src/pages/profile/default/components/ranking/index.ts
@@ -1,0 +1,1 @@
+export { ProfileRanking } from './Ranking'
