+++ src/pages/profile/default/components/user-profile/index.ts
@@ -1,0 +1,1 @@
+export { UserProfile } from './UserProfile'
