+++ src/pages/profile/index.ts
@@ -1,0 +1,2 @@
+export * from './default'
+export * from './profile-question'
