+++ src/pages/profile/profile-question/components/answer-box/index.ts
@@ -1,0 +1,1 @@
+export { AnswerBox } from './AnswerBox'
