+++ src/pages/profile/profile-question/components/index.ts
@@ -1,0 +1,3 @@
+export * from './answer-box'
+export * from './question-box'
+export * from './write-reply'
