+++ src/pages/profile/profile-question/components/question-box/index.ts
@@ -1,0 +1,1 @@
+export { QuestionBox } from './QuestionBox'
