+++ src/pages/profile/profile-question/components/write-reply/index.ts
@@ -1,0 +1,1 @@
+export { WriteReply } from './WriteReply'
