+++ src/pages/profile/profile-question/index.ts
@@ -1,0 +1,1 @@
+export { default as ProfileQuestionPage } from './ProfileQuestionPage'
