+++ src/pages/profile/profile-question/utils/index.ts
@@ -1,0 +1,1 @@
+export { default as formatDate } from './formatDate'
